Updated failing and missing tests
